#8 - Fixed beam relaxation case, by making the orthogonal beam velocity the sum of x and y. This case is not exactly as in the paper

+++ pysrc/simulation/coulomb_collisions/collision_models/abe_collison_model.py
@@ -176,7 +176,7 @@ class AbeCoulombCollisionModel(object):
             assert vel.shape[0] == self.__n_1 + self.__n_2
         assert vel.shape[1] == 3
 
-        num_steps = math.ceil(final_time / dt) + 1
+        num_steps = int(math.ceil(final_time / dt) + 1)
         vel_results = np.zeros((vel.shape[0], vel.shape[1], num_steps))
         vel_results[:, :, 0] = vel
         times = np.zeros((num_steps,))
